anchorage: track the high watermark

+++ runtime/include/alaska/service/anchorage.h
@@ -8,7 +8,7 @@ extern "C" {
 
 struct anchorage_metadata {
   int locks;
-  uint8_t flags;
+  uint32_t flags;
 #define ANCHORAGE_FLAG_LAZY_FREE (1 << 0)
   uint64_t last_access_time;
 };
